Got a working example fade-out pre-filter.  Woot.
I realized my earlier assumptions about a pre-filter were incomplete.  They need to be within the context of a sequence because they have timing relative to the sequence.  They need a start time, a time span, and target nodes and start to look a lot like our other *Node types.  So I created an IDataNode interface, changed EffectStream to DataStream, EffectStreams to DataStreams, and included a stream for pre-filter nodes so that the sequence could easily store and retrieve them.

The pre-filters don't change during execution of a sequence.  So the context builds a cache of what channels get what filters.  When the ProgramContext changes sequences, it rebuilds the cache based on the filters in the current sequence.

I tested this by hard-coding a 5-second fade-out filter that's applied to all nodes (and therefore all channels).  I put in a ramp effect that stayed at 100% and a fade-in ramp.  Both started within the filter's time span and exceeded the filter.  The constant ramp did fade out and the fade-in ramp was pretty muted.  After the filter expired, both popped back to 100% of their potential.

Also played around with a 1024-channel rig that has a 30-second ramp up progressively over the first few-hundred channels.  Hmm, I didn't run that one with the pre-filter.  (...time passes...)  Well it seems to work and didn't seem to lag.  But I'm also tired.  The results weren't a screaming disappointment, I'll say that.

Most of the changes are the result of name refactoring.
Post-filters are next, but they will have to wait until my next day off.