A quick-and-dirty but operational post filter that converts color commands to grayscale (monochrome commands).  It is a little clunky to use because you have to reconfigure the dummy controller when you want to change between monochrome and polychrome.  But I had an effect that did a gradient from blue to yellow.  In color mode, the color effect using the color intent resulted in a good gradient.  When the Grayscale filter was applied (I hard-coded it to be applied to the first output of every controller) and reconfigured the dummy controller for monochrome, it resulted in a successful dark-to-light gradient.  Took it back to color mode, left the filter intact, and there was no output, as expected (monochrome commands versus the color command the controller expected).
There is some project cruft resulting from this as I bopped around with the existing effects when trying to come up with a color intent and effect.  But hey, it's a branch, right?  It's all a dream.  What was used was Grayscale (post-filter), ColorIntent (intent), and ColorWithIntent (effect).
Of the changed files, the only real changes were to ProgramContext (removed the hard-coded pre-filter), OutputController (passing post-filter methods through to the outputs), and VixenSystem (added hard-coded post-filter).  Other changes were "damn fool, why did you do that" or other such minors.

I'm excited to move it off my Beowulf cluster to see how it works on average hardware. ;)  Nah, not that bad, but I do need to fire up my old desktop and see how this fares.  So far, I'm pleased.  But I may also be deluding myself.