This will break everything you ever loved.
A lot of files in this commit aren't actually used.  They were for a time and then abandoned, but not deleted "just in case".  But it's still a huge frickin commit.

Got it compiling and running again.  Nothing has been put into namespaces yet.  This version carries snapshots of intents and filters down to the controller and lets the controller do the actual data generation via an output module-specified data policy.  Hacked together a couple of intents and data values, along with the fade-out pre-filter and a color-splitting post-filter.  It's a step.

Among the cool side effects of these changes is that you can mix what kind of data you send not only to a controller, but even an output.  For example, output 1 gets non-color data while output 2 gets color data AND non-color data, and it just works.

After doing some testing of dynamic dispatching using the 'dynamic' keyword versus double-dispatching, the latter won by a factor of 3.5.  So there is a ton of double-dispatching done in here to allow for dispatching based on an object type at runtime instead of what it is declared as at compile time.  I'm new to this, but it's starting to make sense.  For example, an IFoo gets handled as an IFoo<T> without any casting or use of the 'is' operator.

Reimplemented serialization of all file types.  That whole namespace is a mess right now.  Basically, every file has a driver class that is decoupled from how it's stored.  There are a bunch of ancillary classes in there that were fun at the time, but time will tell if they will stay.

Oh yeah...Sleep() and our frame rates.  I changed the instrumentation values a bit.  It now reports the requested sleep time for a controller thread and how long it actually sleeps.  I could not get it to go above 32 Hz when specifying 50 Hz.  Turns out Windows 7 has a timing resolution of 15 ms when sleeping (among other things), so controller update frequencies kept coming in at intervals of 15 ms.  But this hasn't happened all the time, I don't know why.  Anyway...I found something that lets the application register as a multimedia application, as if it were playing video or audio, and then unregister when it closes.  The result is that our sleep resolution went down to 10 ms and our thread priority goes up.  The sleep times are now more in line with what's being requested.
This only works for Windows Vista and above.  For XP, there is an API call that can be made, and is, but it's not something MS likes to have used, so it may go away.

Haven't tried any high channel counts yet.  Too dang tired.  Good night.