90% of the changes are files moving to a different folder, but I'm too tired to do an SVN fixup to maintain the history.  The files were new with the previous commit anyway and they may not remain; nothing core affected by the moves, just some presentation controls.
Got the filter rule working; takes a patching rule so that the filters are applied to the same outputs that are patched.
Created a filter template file type for that.
Patching rule now breaks up its results by channel so that the patching and filter rules will affect the channels the same, instead of whatever is consuming the rules doing it.
Ran the dummy controller with data for %, floats, longs, and color to test each (separately).  All worked, eventually.
Eyes fell from their sockets; replaced them after a quick soak in coffee.

All presentation is crap; just enough for concept testing.

Ugg tired, go sleep now.  Wake when sun in sky.