Lawdy, lawdy, lawdy...where does one start?
Experimenting with the concept of making controllers part of an "output device" class and a sibling to a "preview" output device.  With that comes "output" modules becoming "controller" modules and a new "preview" module type.  This allows previews to get the same treatment as controllers including the specification of a data policy to control how its data is generated.

Previews get their very own UI thread.  Since previews are expected to be UI (as opposed to controllers, which can be but aren't expected to be) and could be burdensome, I'm experimenting with moving them to their own UI thread so that the editors and admin app will not be affected by them.  It's been composed in a manner that allows the user to move them back to the standard UI thread (SingleThreadBehavior.cs, MultiThreadBehavior.cs).  They also don't need to be Winforms forms.  There are overrides to allow the use of an ApplicationContext.

Playing around with compound effects and combining them using boolean operations.

Refactoring the data policy classes to use base classes to clear out some boilerplate code and to provide future system-enforced behavior.

Added a couple of system config values to disable filters and compound effects.  I haven't benchmarked it yet, but the expectation is that it will lessen the runtime load, should someone need to do that.

Patching and filter tool fixes.

The ChannelNodeService makes use of a ChannelPatchingChangeSet to buffer channel changes.  This further buffers changes to output patching.  The changes take place when the changesets are committed instead of being applied live.  This is to remove the need to reload the sys config if the user decides to cancel config changes.  However, I haven't changed it to stop doing the reloads yet.  These changesets also allow changes to be limited in scope to what actually changed instead of needing to reload all changes or restart controllers because one channel or output changed.

It's been fun!