I believe my spoken words were "Eff, it works." (The implementation may not yet be ideal, but I couldn't believe it worked.)
So much to say...

A big step towards Michael's intent vision.  It filters intents and caches and buffers them.

The buffer only holds 10 seconds of data right now, but it's configurable.  So far this seems to be adequate.  I added a performance counter to monitor the buffer's level, in seconds (could/should have done %...I will change that) and it holds steady at around 10 seconds, lowering only during the last 10 seconds of execution.

I don't have any stats on the cache yet, but I know it's having an effect because, previously, my 1000-channel, 3000-output rig was causing the dummy lighting module to occasionally request a negative sleep value, but it doesn't any more.

Dispatching on intent segments to avoid ugly conditional blocks.  Not as fast as a cast, but my profiling so far doesn't show it to be a big difference.

Some movement towards interfaces.

The placement of the cache and buffer are:

Context -> Executor -> Buffer -> Cache -> Sequence data

Why: The read-ahead buffer will pull data through the cache and buffer pre-filtered intents.  Then the context pulls from the buffer as it needs.

The cache needs to persist, so it couldn't be part of the data source since the data source is re-instantiated with every execution.  The context and executor persist until the context is released by the editor, so the executor owns the cache (and uses it as the IDataSource for the context).

The buffer needs to be cleared with every execution.  So it would have been nice to have the cache close to/with the executor and context, since they all need to persist, and the buffer be closer to/part of the data source, since they need to be cleared/re-instantiated with every execution.  However, that would mean that the cache couldn't pre-load with filtered intents because data would only pass through it when the context requested it, which defeats the purpose of doing all of this.  So the executor owns the cache and provides it to the buffer as its source of data.  The cache passes the requests through to the actual data source, as needed.

Just happy it works!