Post-Filters: Bringing post-filters back into the fold.  I've only tested with the color filter so far.  Probably the biggest change to the post-filters is that they no longer operate upon value types but upon the command objects that wrap the value types.  This allows a filter to take in one command type and output another.  Now the color filter doesn't have to do something kludge like take in a color and cram the intended byte value into the B component.  It can take in a color value and output a byte value.  Likewise, if the filter was going to convert a short to a byte, it can take in a short command and output a byte command.
Generator, Combinators:
Since post-filters take in commands, they needed to go after the data policy had generated a command.  However, the filter may be crucial in generating that final command.  For example, the color filter...a generator would generate a byte command, which can no longer be filtered by the color filter because it's lost the color information.  But the filter still has to affect the final value for the output, so it has to happen after everything else.  The result: Generators are gone.  The data policy's combinator now produces a command instance instead of a value type.  That command goes through the filters and onto the output as its state (assuming the command makes it through the filters, it may not).

The application of channel patches and filter templates can now add to what's there or replace it.

More support for the "lighting" value type and intent.

In addition to the controller's data policy, an output can override it.  No way to do that through the UI currently, but the mechanism is there.  Very light impact.

Combinators are ugly right now, but I don't see another way around it yet.  Details are in Combinator.cs.

Added pause and resume to controllers.  The underlying module supported them, but the controller didn't.  Now the application of channel patches and filter templates can be done safely by pausing and resuming the affected controllers (which PostFilterService does for you).

I'm seeing awesome performance now with the cache and pre-filtering of effects.  Not fully tested yet, but working well.