(I intended to commit before it grew this large, but time got away from me this week.)
Defined an initial "smart controller" module type.  The differences between this type and the usual snapshot controller are that it takes intents instead of commands and it doesn't have a data policy since it doesn't do anything with commands.  I pulled the IDataPolicy reference out of the underpinning output device code and moved it down the hierarchy to the controller and preview implementations.  Haven't created the SmartOutputController class yet, but that's next.

More use of interfaces over concrete classes.
More comment clean-out.

I pulled out the IntentState concrete sub-classes and just made it *the* class.  Could have done the same with the Intent sub-classes, but I'm keeping them around because they can enforce type-specific boundaries for the intent types.  For example, the double intent type is used to implement percentage values and restricts the values to being between 0 and 1.

Added Pause and Resume methods to OutputDeviceManagerBase as well as a supporting method or two.  Useful when patching or changing the output filters of a number of outputs and/or controllers.