Smart controller changes.
Since the new smart controller type and the existing controller type share so much in their module handling, I pulled out a superclass of ModuleBasedController.  There's still a lot of commented-out code in there that I will clean out after this commit.

I'm doing a lot of subclassing that I may regret.  But the two controllers have pretty much the same behaviors, only the output and data types are different.  And a bit of the data handling.