Added three new modules -- 1. Grid property 2. ImageGrid effect 3. FilePath-type editor
They're rudimentary right now, but it works.  You can define (through the Grid property) a node as having a collection of channels that are to be treated as a grid.  Then you use the ImageGrid effect on that wrapping node.  Configure it to select an image file to display and you're done.  Nothing fancy yet, no fades or animations, just a static image.

Word to the wise: If you're going to do this with a 100x100 grid, DO NOT try this in the timed sequence editor.  10,000 nodes is too much!  Just script it.