Changed output filters to work against intent states instead of commands.  They will affect the intent state before it goes through the controller's data policy.  This is because once it is evaluated by the first step of the data policy, the intent is lost so the filter won't know a color from a hole in the ground.
Created a StaticIntentState class because the current IntentState class pulls a live value from the intent.  The filters need something that will allow the state value to be set.

Also renamed IIntentStateList.cs to IIntentStates.cs, which had some cascade effects (which is most of this changeset).

None of this is tested yet.