This is the move towards trying out Michael's filter model.  What else...umm...
Updates requested by a controller now happen in three phases:
 1. Channels (from contexts)
 2. Filters (from channels and each other)
 3. Outputs (from whatever they're connected to, a channel or a filter)

For future reference:
 Data flow model = Channel layer -> Filter layer -> Output layer
 Channels, output filters, and controller outputs are all participants in this model.

There is now VixenSystem.DataFlow.  It has methods to add, remove, and connect participants in this new data flow model.  Adding/removing channels or outputs will cause participants to be added or removed.  But to add output filters, you do it through VixenSystem.DataFlow.  This is because channels and outputs have other primary responsibilities and are only voluntarily participating in this model.  Output filters, however, are 100% dependent upon it.  Also, using the intended workflow, the user will be dropping filters into a workspace and connecting filters, outputs, and channels as they please.  I thought it would be easier if the app facilitating this only had to go through VixenSystem.DataFlow to accomplish all of this instead of thinking "it's a filter, so I go through VixenSystem.Filters...it's not a filter, so I go through VixenSystem.DataFlow".  Just always do the latter.

No more patching and no more output filters in the sense that they are filters attached to outputs.  Instead, they are filters for data destined for outputs.

Pulled any system-level module data out of SystemConfig.  A non-static dataset was added to ModuleStore and system-level objects are redirecting their local dataset references to that.  Currently, that is used by controllers, properties, and output filters.  These are the module types that are system-level and need instance data, not static type data.

It used to be:
Channel -> (intent states) -> Evaluation -> (values) -> Combination -> (command)
It is now (assuming no filters):
Channel -> (intent states) -> Evaluation -> (commands) -> Combination -> (command)

Gone are fun classes like LightingCommand, in their place are stoic, serious classes like _8BitCommand.

The DummyLighting controller now changes its data policy based on its configuration because of this.  Different commands need to be generated by the policy.  This does make more sense than a single data policy trying to be responsible for any mode of execution.

Much of this is not tested, many modules are not compiling, but the core framework runs.
A lot of crap is commented out; I will be cleaning that up.
I will also be getting the existing output filters refactored to conform to their new roles.
Don't bother trying to execute any effects because with the lack of patching and no GUI facility for creating a path from a channel to an output, the data will go nowhere.

I can hear Michael screaming at his monitor, "ONE COMMIT?!  YOU ARSE!  WHY DID YOU...WHY!!  BAAAH!!" :)  Because this change encompassed so many changes, I wanted to commit them wholesale so there isn't any "in which commit was X changed?  (search, search)"  Just go back to this commit, find the file, see the change.

That is all.