Trying to bring back some patching rules to facilitate easier patching.  Just starting with one whose purpose in life is to take in some channels, a controller, and a starting output on the controller and generate source references for the outputs or full patch instances.
No actual use of this yet, but its presence has no side effects.
UI still needs to implement a specific control to enable the user to configure the parameters for the rule.  Then the UI can invoke the rule, generate the patches and either show them to the user or use them to set sources.  For example, ChannelsToSingleController would generate patches whose source references a channel and targets an output.

I don't know if there's any use for the GenerateSourceReferences member of the interface since the references don't give you *what* is doing the referencing.  Maybe there are simple patching use cases where this makes sense?  If not, we can drop it.