When adding, removing, or enumerating from a DataStream, sort the data first to ensure it's in StartTime order.
This is only a bit of a dodgy hack at the moment to get things working; it won't properly work with LiveListEnumerator
if we want to add/remove data from the list on the fly. We probably need to stop using a LiveListEnumerator, pending
what KC thinks about how it all works.