Added support to Color Gradients and the associated controls to give it the 'concept' of discrete colors:
The gradient itself doesn't store whether the gradient should be discrete colored or not: that is a property of the element it's applied to. So, all the data in the gradient is stored the same way: the convention is now that if multiple colors are enabled at a particular point, there are multiple ColorPoints on the line, one for each color, all at the same position.

The controls that display it to the user are what is used to show it is discrete: it renders differently if so.  Also, the editor used is different: it only allows the user to pick one or more of a set of colors.

It's the editor's responsibility to determine if the effect it is editing is applied to a discrete colored item; and if so, ask the gradient to render appropriately.