Changed the listbox event from selected index to mouse click.  Allos me to clear out the selected item in the listbox and reset the mappings.
corrected the problem with the colors.  I know longer look for the color by name, I set the back color using the argb values and if there is a name it is shown.

When the user presses ok/cancel in the mapping form the convertef form will clear the map text box, disable the convert button until an item is selected from the listbox.

added double buffering to the forms.