Here's my shot at the new StaticLightingArrayIntent.  This intent is used by the (updated) nutcracker effect and could also be used by an 'animation'/movie effect in the future.  I verified it works (via Giz's sequence) up through the Preview Display.  We need a volunteer (Giz?) to verify it works OK going to lights.
This version is not quite as memory efficient at the previous 'NutcrackerIntent' because it stores LighingValues which take up 25$ more space than the previous version with stored Colors.  This means instead of a 10X reduction in memory (on the Giz sequuence) this one is only 7X better - still... not bad.