On my system the preview 'shimmers'/'flickers'.  I wondered if it might benefit from double buffering, so I took a shot.  Unfortunately, I don't know anything about DX, so there's probably a way better way to do this.. I've seen mention of an hwnd render target that is always double buffered and also some DX3D 'swap chain' thingee... I didn't know how to do either of them, so I took the dumb way out.
This change simply creates a local, temp, CompatibleRenderTarget that is a bitmap.  Most of the rendering uses this local (rt).  When just about done we get the bitmap out of rt and draw it on the real render target.  On my system the preview looks nice and smooth and seems only a little slower (maybe 10%, but the stats flickered too much before to really be sure).

I also tweaked the stats reporting to be a little more self contained and to make the average and max use about a 5 second window (was 50 sec).

Let me know what you think?

PS: I added a 'using' so the diff looks worse than it really is if you are looking at the GitHub diff (Tortoise shows id fairly well).