Performance tweaks uncovered by Giz.  This version is a little more surgical than Giz's (no changes to unrelated files).
The DefaultValueArrayMember.cs change takes advantage of the fact that the PropertyInfo array being computed is really the same for all instances of a given class. We simply keep a cache of arrays keyed on owner.GetType().  This change dropped the render time for the Giz3 sequences from 146 sec to 102 sec.

The ScheduleItem.cs change is straight from Giz. Simply put is runs the render of the sequences in parallel (if you have multicores).  It drops the Giz3 render time from 102 sec to 85 sec.  My 4C/8T cpu was not maxed out doing this... can only guess we have some hot locks somewhere preventing more parallelism.