Enhancement for the web server module.
Reorganize into a framework that is more receptive to adding more features and expose all API that the web front end uses for other outside applications to take advantage of.

Refactor the original element code into the new framework and expose all the API for consumption. Make the mobile web page consume that UI to do it's functions.

Add ability to provide the color that full color elements are turned on with instead of defaulting to white.

Add the ability to play sequences that reside in the home sequence directory. Sequences can be played and stopped. Provide a status to get the current playback position on demand of the current test sequence playing. Full API used is exposed.