Nutcracker like to have a rectangular set of elements to work on - a collection of groups that have elements. If the user applies Nutcracker to an irregular shape of groups/elements bad things happen.  Greg S. has a sequence like this.  A Nutcracker would not render in the Grid and would crash if the Effect editor was invoked.
The root cause was weakness in the StringCount and PixelPerString calculators - and the fact that there are two (different) copies of this code (one in NutcrackerEffectEditor/NutcrackerTypeEditorControl.cs and one in Nutcracker.cs).  This change fixes the same defect in both of them and brings them into alignment to behave the same.