Clean up of previous commit. Makes better use of available methods
The Dialog now determines how many effects are possible for addition and displays
this in a label at the top of the form.
The form caption also shows which effect is being worked with.

If something goes wrong with calculations the user is prompted with a Yes/No continue anyway option.
This takes place once the user clicks OK on the dialog.