This work is for VIX-471 & VIX-472
Adds Drawing mode to sequence editor, which allows users to draw effects into the grid vs drag and drop.
Users can select this mode as an option.

Adds indicator lines to effects that extend top to bottom of the grid when resizing or dragging an effect.
This is an option on the Edit menu, user can select color.

Also removes a stray ; in the if statement for resuming play in the editor using the space bar, that would
cause pass thru of else and stop the sequence