Found small issue that when using the mark manager to paste effects to beat marks, if the active row was selected rather than highlighted, all of the created effects would be stacked up on top of each other.
The fix is to determine at the time of the call weather or not the active row is selected, and if so pass a time of zero.
Previously this check was done in the method that does the actual pasting, and is what allowed the other items that use the method to stack effects up when not meaning to.

This also would have negative results when using the lipsync to align to the beat marks, no changes were needed other than to the paste method, and the menu item
which makes the call.