VIX-664: Varies slightly from the ticket description. When multiple types of effects are selected for edit, you now get a sub-menu that lists the types, you click the type of effect you intend to edit. Allows easier editing in some situations.
VIX-688: Clean-up of contextMenu this is the initial work on this ticket. It gives all of the contextmenu items a name is that of a consistent format.

VIX-674: This fixes a minor issue where the Cut/Copy/Paste/Delete menu items were being deleted from the Edit menu after the contextmenu was accessed.