E1.31 Plugin Overhaul
Reworking the E1.31 plugin quite a bit. Each instance of the controller
now has a single destination. This should make it easier to setup the
plugin and reinforces some good workflow habits for the plugin. The
start value now auto-populates and is managed auto-magically by default.
The TTL column has been removed and hard-coded.  The setup form has been
divided with a tabbed interface to hide the more advanced options that
95% of people don't know or care about. The plugin now supports
transmitting blind data and a user-defined priority. The plugin also
assigns a name to its outputs, there is another commit following that
updates the display setup form to show the plugin-defined names.

Known bugs:
Checking for universe duplicates/conflicts needs to be reviewed and
double-checked.
Universes with 5 or less channels aren't transmitting correctly. I
checked older versions of the plugin and this isn't a new bug.

The updated plugin will load settings from older versions as best it can
but some repatching may need to be done depending on each users setup.