VIX-797 Fix up a whole slew of bugs in the Marks Dock window. The docker itself was not being correctly managed when it is opened/closed. Thus it was partially initialized at times and had bad state or missing event handlers. Reworked how it is created to make sure it is properly initialized. Also cleaned up how it is closed. This fixes the following issues.
Crash when trying to add new marks after the window has been closed.
Marks not being properly removed from the timeline after they are deleted in the Marks docker window.
The sequence not being set to changed when marks are turned on or off.
Clean up of event handlers to avoid memory leaks.