Add shortcut keys for the alignment functions in the editor based on a user request. These will work when more than one element is selected in the time line and the mouse is hovered over one of the selected elements.
S         Align Start times of the selected elements to the key element under the hover.
S + Shift Align Start times of the selected elements holding the duration to the key element under the hover.
E         Align End times of the selected elements to the key element under the hover.
E + Shift Align End times of the selected elements holding the duration to the key element under the hover.
B         Align both start and end times of the selected elements to the key element under the hover.