Big rework of the audio effects to bring them into mainstream. Added some core support for effects to support media. This allows them to specify that they can use the current media associated with the sequence. When they do that a reference to the media will be provided. They also get the ability to know their start time. This allows them to offset into any media if needed.
Next the audio effects have been reworked to use the new media support. This replaces some kludgy code that was needed to find the audio. Enhance them to support discrete colors correctly. Before there was no real support for discrete and they could just happen to work in places if the colors hit right. This allows some better uses of them on some discrete props. For the most part these will be used on pixel items. The depth of effect feature was added to help define which level things need to act on. This gives more flexibility on various props to determine how they work.

Still need to do some work on the property and categories order in the editor, but that should be minor work after this.