VIX-2430 Move SVG library to the core libraries. Remove it from the Vixen core project. Remove the copy local reference in the Shape effect so it will use the library version. The Shape project was not building the x86 version.
Clean up some Mark copy local references. 

Upgrade the web server project dependencies as they are several versions behind. Owin, SignalR and supporting libs.

Upgrade the Docker library to the current version. 

The ColorWash effect was not build with optimization on in the release build config.

Remove a shortcut link to the Debug output that does not belong.

 