VIX-2478 Add some logic to better handle video files. Stream the video to a audio less video file that has been size, scaled, rotated, etc that then can have the images picked off one at a time at render time. This greatly reduces the space needed on disk to store the image frames. It does help a bit with performance due to the reduced file access.
TODO this needs the 32 bit version worked out. The are two separate nuget packages. The logic that is used by querying the output of the ffmpeg to determine options for the conversion might be able to be picked up without the first call out to ffmpeg. So see if that can be accomplished ot help out this performance some.